Can now set IMDB rating when adding a movie.

+++ src/BLL/MovieManager.java
@@ -22,7 +22,7 @@ public class MovieManager {
         return movieDAO.getMovieById(id, categories);
     }
 
-    public Movie createMovie(String title, String filepath, java.sql.Date lastWatched, int personalRating, int IMDBRating) throws Exception{
+    public Movie createMovie(String title, String filepath, java.sql.Date lastWatched, int personalRating, float IMDBRating) throws Exception{
         return movieDAO.createMovie(title, filepath, lastWatched, personalRating, IMDBRating);
     }
